+++ objectified-ui/public/WHATS_NEW.md
@@ -80,6 +80,8 @@ We continue to improve the platform based on your feedback with improvements and
   - Added response header templates to the response properties form
   - Dragging a property to a variable chip in the header will apply that property's primitive settings to the bound variable
   - Dragging a class to a response schema allows the class to be copied or referenced
+- Control Panel:
+  - Versions page now has the ability to view the relationship graph for each version
 - OpenAPI Support:
   - Now at 100% support for class and properties definitions
   - Improved import support for paths and securitySchemes